[CHORES] Moved tests/map/MapPage test into tests/app/page
